Moved home key func and popup
